Prep backend file locking and dark mode
Dark mode CSS and script for the home page are staged and commented out.

import fcntl has been added to support file locking but it remains in testing.

Should this support Windows as well? I'm hesitant as servers are typically Unix based and I see no need for this on Windows Server